+++ src/UIComponents/Status/index.js
@@ -1,0 +1,2 @@
+import Component from "./Status";
+export default Component;