+++ CHANGELOG.md
@@ -14,3 +14,4 @@
 - Added `caret` in Select
 - Fixed Select dropdown padding
 - Select dropdown overflow on select
+- Added `Status` component